+++ plc4go/internal/plc4go/bacnetip/readwrite/model/BACnetServiceAckReadProperty.go
@@ -394,7 +394,7 @@ func (m BACnetServiceAckReadProperty) Box(name string, width int) utils.AsciiBox
 	childBoxer := func() []utils.AsciiBox {
 		boxes := make([]utils.AsciiBox, 0)
 		// Const Field (objectIdentifierHeader)
-		boxes = append(boxes, utils.BoxAnything("ObjectIdentifierHeader", 0x0C, -1))
+		boxes = append(boxes, utils.BoxAnything("ObjectIdentifierHeader", uint8(0x0C), -1))
 		// Simple field (case simple)
 		// uint16 can be boxed as anything with the least amount of space
 		boxes = append(boxes, utils.BoxAnything("ObjectType", m.ObjectType, -1))
@@ -402,7 +402,7 @@ func (m BACnetServiceAckReadProperty) Box(name string, width int) utils.AsciiBox
 		// uint32 can be boxed as anything with the least amount of space
 		boxes = append(boxes, utils.BoxAnything("ObjectInstanceNumber", m.ObjectInstanceNumber, -1))
 		// Const Field (propertyIdentifierHeader)
-		boxes = append(boxes, utils.BoxAnything("PropertyIdentifierHeader", 0x03, -1))
+		boxes = append(boxes, utils.BoxAnything("PropertyIdentifierHeader", uint8(0x03), -1))
 		// Simple field (case simple)
 		// uint8 can be boxed as anything with the least amount of space
 		boxes = append(boxes, utils.BoxAnything("PropertyIdentifierLength", m.PropertyIdentifierLength, -1))
@@ -416,11 +416,11 @@ func (m BACnetServiceAckReadProperty) Box(name string, width int) utils.AsciiBox
 			boxes = append(boxes, utils.BoxBox("PropertyIdentifier", utils.AlignBoxes(arrayBoxes, width-4), 0))
 		}
 		// Const Field (openingTag)
-		boxes = append(boxes, utils.BoxAnything("OpeningTag", 0x3E, -1))
+		boxes = append(boxes, utils.BoxAnything("OpeningTag", uint8(0x3E), -1))
 		// Complex field (case complex)
 		boxes = append(boxes, m.Value.Box("value", width-2))
 		// Const Field (closingTag)
-		boxes = append(boxes, utils.BoxAnything("ClosingTag", 0x3F, -1))
+		boxes = append(boxes, utils.BoxAnything("ClosingTag", uint8(0x3F), -1))
 		return boxes
 	}
 	return m.Parent.BoxParent(boxName, width, childBoxer)
